+++ CHANGELOG.md
@@ -7,7 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 
 ### [Beam]
+
 - Refactored beam `CompressAggregations` command to run in chunks because of colliding database transactions with `AggregatePageviewLoadJob` command, which caused deadlock. remp/remp#944
+
+### [Mailer]
+
+- Fixed broken new email page when no layout or newsletter list was defined.
+- Fixed broken new email template page when no layout or newsletter list was defined.
 
 ## [0.27.1] - 2021-07-08